ui done 	modified:   lib/data_layer/mempool.dart 	modified:   lib/data_layer/robot_database.dart 	modified:   lib/data_layer/tx_database.dart 	modified:   lib/domain_layer/account.dart 	modified:   lib/domain_layer/operation.dart 	modified:   lib/domain_layer/transaction.dart 	modified:   lib/presentation_layer/common_layout.dart 	modified:   lib/presentation_layer/constants.dart 	modified:   lib/presentation_layer/screens/account_screen.dart 	new file:   lib/presentation_layer/screens/confirmed_operations_screen.dart 	modified:   lib/presentation_layer/screens/my_robots_screen.dart 	new file:   lib/presentation_layer/screens/pending_operations_screen.dart 	modified:   lib/presentation_layer/screens/registration/sign_in_screen.dart 	modified:   lib/presentation_layer/screens/registration/sign_up_screen.dart 	modified:   lib/presentation_layer/screens/transfer_rights_screen.dart 	deleted:    lib/presentation_layer/widgets/empty_banner.dart 	new file:   lib/presentation_layer/widgets/empty_banners.dart 	modified:   lib/presentation_layer/widgets/key_card.dart 	modified:   lib/presentation_layer/widgets/loading_indicator.dart 	modified:   lib/presentation_layer/widgets/receiver_picker.dart 	modified:   lib/presentation_layer/widgets/robot_card.dart 	modified:   lib/presentation_layer/widgets/robot_picker.dart 	new file:   lib/presentation_layer/widgets/transactions_table.dart 	modified:   pubspec.lock 	modified:   pubspec.yaml 	modified:   server_example/bin/blockchain/account.dart 	modified:   server_example/bin/blockchain/transaction.dart 	modified:   server_example/bin/database/mempool.dart 	modified:   server_example/bin/database/robot_database.dart 	modified:   server_example/bin/database/tx_database.dart 	modified:   server_example/bin/server.dart

+++ server_example/bin/blockchain/account.dart
@@ -75,8 +75,9 @@ class Account {
     return true;
   }
 
-  /// Adds the given robot to the corresponding account in [RobotDatabase].
-  Future<void> addRobot(Robot robot) async {
+  /// Adds the given robot to the account.
+  /// If `updateDB` is false, account does not get updated in [RobotDatabase].
+  Future<void> addRobot(Robot robot, {bool updateDB = true}) async {
     // ensuring that the given robot doesn't belong to the account
     if (robots.contains(robot)) {
       throw ArgumentError(
@@ -86,7 +87,7 @@ class Account {
     }
     try {
       // adding a robot in robotDatabase
-      await blockchain!.robotDatabase.addRobot(robot);
+      if (updateDB) await blockchain!.robotDatabase.addRobot(robot);
       // updating current state
       robots.add(robot);
     } catch (e) {
@@ -94,15 +95,16 @@ class Account {
     }
   }
 
-  /// Removes the given robot from the corresponding account in [RobotDatabase].
-  Future<void> removeRobot(Robot robot) async {
+  /// Removes the given robot from the account.
+  /// If `updateDB` is false, account does not get updated in [RobotDatabase].
+  Future<void> removeRobot(Robot robot, {bool updateDB = true}) async {
     // ensuring that the given robot belongs to the account
     if (!robots.contains(robot)) {
       throw ArgumentError("The Robot is abscent in {robots}", robot.toString());
     }
     try {
       // removing a robot from robotDatabase
-      await blockchain!.robotDatabase.removeRobot(robot);
+      if (updateDB) await blockchain!.robotDatabase.removeRobot(robot);
       // updating current state
       robots.remove(robot);
     } catch (e) {
@@ -130,6 +132,16 @@ class Account {
       robotID,
     );
     return operation;
+  }
+
+  /// Returns all confirmed operations, which were performed by the account.
+  Future<List<Operation>> getPendingOperations() async {
+    return await blockchain!.mempool.getAccountOperations(accountID);
+  }
+
+  /// Returns all confirmed operations, which were performed by the account.
+  Future<List<Operation>> getConfirmedOperations() async {
+    return await blockchain!.txDatabase.getAccountOperations(accountID);
   }
 
   // equivalent of printBalance()
